Update Issue Implemented.

+++ controller/financialDetails.controller.js
@@ -255,65 +255,29 @@ exports.updateFinacialDetail = async (req, res) => {
         req.files.financial_data?.length > 0 ||
         req.files.financial_data2?.length > 0
       ) {
-        if (
-          (fDetails.financial_data === req.files.financial_data
-            ? req.files.financial_data[0].path
-            : "") ||
-          (fDetails.financial_data2 === req.files.financial_data2
-            ? req.files.financial_data2[0].path
-            : "")
-        ) {
-          const financial_data = financial_data_DocPath;
-          const financial_data2 = financial_data2_DocPath;
-          req.body.financial_data = financial_data;
-          req.body.financial_data2 = financial_data2;
-          FdetailSchema.update(req.body, {
-            where: { userId },
-          })
-            .then(() => {
-              res.status(200).send({
-                message: "Financialdetail was updated successfully!",
-                status: "success",
-              });
-            })
-            .catch((err) => {
-              res.status(500).send({
-                message:
-                  err.message ||
-                  "Some error occurred while updating the Financialdetail schema.",
-              });
-            });
+        let financial_data = financial_data_DocPath;
+        let financial_data2 = financial_data2_DocPath;
+
+        if (fDetails.financial_data === req.body.financial_data) {
+          financial_data = req.body.financial_data;
         } else {
-          const financial_data = financial_data_DocPath;
-          const financial_data2 = financial_data2_DocPath;
-          req.body.financial_data = financial_data;
-          req.body.financial_data2 = financial_data2;
-          FdetailSchema.update(req.body, {
-            where: { userId },
-          })
-            .then(() => {
-              res.status(200).send({
-                message: "Financial was updated successfully!",
-                status: "success",
-              });
-            })
-            .catch((err) => {
-              res.status(500).send({
-                message:
-                  err.message ||
-                  "Some error occurred while updating the Financialdetail schema.",
-              });
-            });
+          financial_data = financial_data_DocPath;
           directoryFiananceOneDelete = fDetails.financial_data;
-          directoryFiananceTwoDelete = fDetails.financial_data2;
-          if (directoryFiananceOneDelete) {
+          if (directoryFiananceOneDelete && !req.body.financial_data) {
             fs.unlink(directoryFiananceOneDelete, (err) => {
               if (err) {
                 throw err;
               }
             });
           }
-          if (directoryFiananceTwoDelete) {
+        }
+
+        if (fDetails.financial_data2 === req.body.financial_data2) {
+          financial_data2 = req.body.financial_data2;
+        } else {
+          financial_data2 = financial_data2_DocPath;
+          directoryFiananceTwoDelete = fDetails.financial_data2;
+          if (directoryFiananceTwoDelete && !req.body.financial_data2) {
             fs.unlink(directoryFiananceTwoDelete, (err) => {
               if (err) {
                 throw err;
@@ -321,25 +285,17 @@ exports.updateFinacialDetail = async (req, res) => {
             });
           }
         }
-      } else {
 
-        let financial_data = financial_data_DocPath;
-        let financial_data2 = financial_data2_DocPath;
-
-        if (financial_data || financial_data2) {
-          req.body.financial_data = financial_data;
-          req.body.financial_data2 = financial_data2;
-        } else {
-          financial_data = req.body.financial_data;
-          financial_data = req.body.financial_data;
-        }
-
+        req.body.financial_data = financial_data;
+        req.body.financial_data2 = financial_data2;
         FdetailSchema.update(req.body, {
-          where: { userId },
+          where: {
+            userId: userId,
+          },
         })
           .then(() => {
             res.status(200).send({
-              message: "Financialdetail was updated successfully!",
+              message: "Financial Detail was updated successfully!",
               status: "success",
             });
           })
@@ -347,26 +303,61 @@ exports.updateFinacialDetail = async (req, res) => {
             res.status(500).send({
               message:
                 err.message ||
-                "Some error occurred while updating the Financialdetail schema.",
+                "Some error occurred while updating the Financial Detail schema.",
             });
           });
+      } else {
+        let financial_data = financial_data_DocPath;
+        let financial_data2 = financial_data2_DocPath;
 
-        directoryFiananceOneDelete = fDetails.financial_data;
-        directoryFiananceTwoDelete = fDetails.financial_data2;
-        if (directoryFiananceOneDelete && !req.body.financial_data) {
-          fs.unlink(directoryFiananceOneDelete, (err) => {
-            if (err) {
-              throw err;
-            }
-          });
+        if (fDetails.financial_data === req.body.financial_data) {
+          financial_data = req.body.financial_data;
+        } else {
+          financial_data = financial_data_DocPath;
+          directoryFiananceOneDelete = fDetails.financial_data;
+          if (directoryFiananceOneDelete && !req.body.financial_data) {
+            fs.unlink(directoryFiananceOneDelete, (err) => {
+              if (err) {
+                throw err;
+              }
+            });
+          }
         }
-        if (directoryFiananceTwoDelete && !req.body.financial_data2) {
-          fs.unlink(directoryFiananceTwoDelete, (err) => {
-            if (err) {
-              throw err;
-            }
-          });
+
+        if (fDetails.financial_data2 === req.body.financial_data2) {
+          financial_data2 = req.body.financial_data2;
+        } else {
+          financial_data2 = financial_data2_DocPath;
+          directoryFiananceTwoDelete = fDetails.financial_data2;
+          if (directoryFiananceTwoDelete && !req.body.financial_data2) {
+            fs.unlink(directoryFiananceTwoDelete, (err) => {
+              if (err) {
+                throw err;
+              }
+            });
+          }
         }
+
+        req.body.financial_data = financial_data;
+        req.body.financial_data2 = financial_data2;
+        FdetailSchema.update(req.body, {
+          where: {
+            userId: userId
+          },
+        })
+          .then(() => {
+            res.status(200).send({
+              message: "Financial Detail was updated successfully!",
+              status: "success",
+            });
+          })
+          .catch((err) => {
+            res.status(500).send({
+              message:
+                err.message ||
+                "Some error occurred while updating the Financial Detail schema.",
+            });
+          });
       }
     }
   });
